fix(download-file): Integrating request
Previously, we were using the http module provided by node that is a very low level module for http
request and the problem with it, is that we were not handling special cases like redirections. Now
I'm integrating the request module from npm that is going to handle these special cases for us. I'm
also removing some tests that are not going to be necessary anymore.

closes #19