chore(demo): Adding a large demo
Now the user can try the module using a large:demo option where a lot of images are going to be
downloaded, just to be sure that it is going to work for that kind of cases. In the future, this
should be covered by our tests but for now I'm going to add this.